Added ability to delete reminders (#81)
* Added reminder cancellation

* Added reminder removal

* Added reminder deletion

+++ src/Commands/Menus/DeleteOnReaction.cs
@@ -1,23 +1,29 @@
-﻿using Disqord;
-using Disqord.Extensions.Interactivity.Menus;
-using System;
+﻿using System;
 using System.Threading.Tasks;
+using Disqord;
+using Disqord.Extensions.Interactivity.Menus;
 
-namespace Espeon.Menus {
+namespace Espeon {
     public class DeleteOnReaction : MenuBase {
+        private readonly Snowflake _userId;
         private readonly Func<Task<IUserMessage>> _messageFunc;
 
-        public DeleteOnReaction(Func<Task<IUserMessage>> messageFunc) {
+        public DeleteOnReaction(Snowflake userId, Func<Task<IUserMessage>> messageFunc) {
+            this._userId = userId;
             this._messageFunc = messageFunc;
-
-            AddButtonAsync(new Button(new LocalEmoji("🚮"), async args => {
-                _ = Message.DeleteAsync();
-                await StopAsync();
-            }));
         }
 
         protected override async Task<IUserMessage> InitialiseAsync() {
             return await this._messageFunc();
+        }
+
+        [Button("🚮")]
+        public Task Delete(ButtonEventArgs args) {
+            if (args.User.Id == this._userId) {
+                _ = Message.DeleteAsync();
+            }
+            
+            return Task.CompletedTask;
         }
     }
 }
